LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/other/double-click-range-selection-in-list-item-expected.txt

First
Second
This is some atomicinline text.

FAIL Double click on one list item should not select more than one list item assert_equals: Selected Text Should be equal to first list item expected "First" but got ""
FAIL Double click on one text item should select only one text item assert_equals: Selected Text Should be equal to atomicinline expected "atomicinline" but got ""
